Update patch set 5
Patch Set 5:

(3 comments)

I think I addressed everything.  Please review, but if it looks right, just +1 for now.

I responded to hashar's comments, changed the namespace IDs, and added a documentation comment.

Note, due to the namespace ID change, any existing Draft pages on Beta (with namespace 110) should be dealt with to avoid inconsistent state.

There were only a dozen or so Draft pages on Beta when I wrote this, so I moved them to mainspace manually.

This is -1'ed, so when it's ready other than that issue, I can do any remaining Beta moves right before the merge.

Patch-set: 5