Update patch set 3
Patch Set 3: Code-Review-1

After looking literally closer, the 16X16 version is not as sharp as the previous one. You can see the difference in the first ow of pixels, the original has actual whites while the new has a sequence of grays. There are more differences. Since the previous 16x16 icon was correct, why not just reusing it in the new one?

You can also see the difference in the 32x32 version, and even the 32x32 version looks like it could get more pure black instead of dark gray in some pixels.

Sorry for coming back with more work, M4tx. Your GCI task is closed but... do you want your favicon to be the one deployed? Imagine how many times this icon is displayed every day. It makes sense that we want to work on the details up to the last pixel.

Also sorry for not having looked that close before. This patch came in a very busy GCI time.  :)

Patch-set: 3
Reviewer: Qgil <465@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1