Cirrus config update
The following wikis join test2wiki and mediawikiwiki and cirrus as primary:
cawiki
enwikisource
foundationwiki
itwiktionary
loginwiki
nostalgiawiki
testwiki
testwikidatawiki

The following wikis get cirrus as a betafeatures:
all wikisources

The following wikis get cirrus as a secondary:
all wikinewses
specieswiki

Wikinewses have a special configuration:
$wmgCirrusSearchPreferRecentDefaultDecayPortion.  We default it to 0 which
is cirrus's default.  We set it to .6 for wikinews to lower to score of
pages that haven't been updated recently.

I don't have any good way to guess how many shards to break enwikinews
into so I just sharded it like cawiki, which is of a similar size.  The
same is true for specieswiki but I cut the size to 2/3 of cawiki.

Finally, I removed the warning that the number of shards for dewiki was
done before deployment.  It has the correct number of shards.

Change-Id: I5f4c2729d9995004df88a9f1c0905aaa54c7fd73