Update patch set 4
Patch Set 4: Code-Review-1

Wow, great improvement!

I'm -1 the patch just yet, because I agree with Quim that it's possible to make the 16px even better than now, and because there is visible difference between the font colour in the 32px and 48px layers -- the 48px version appears more black, whilst the 32px seems to use gray color for font.

If you fix that, I'd be happy to +1 the patch.

Thanks!

Patch-set: 4
Label: Code-Review=-1