Update patch set 3
Patch Set 3:

Just a bit of context for Aquifacae: in Wikimedia each project is quite sovereign. Wikimedia in English can decide one thing, Wikipedia in Japanese can decide another, then Wikisource in Russian... long etc.

Anything can be discussed and good arguments are always heard. However, usually the "political" discussions happen in the projects, while here in the MediaWiki / Wikimedia tech context we try to keep the discussions technical. This is why in this Gerrit change we just improve the current favicon without challenging the initial design. It is good to have around senior contributors like Odder and Nemo who not only know the technical aspects, but they also can explain why things are the way they are, and point you to the past and current discussions and decisions about these matters.  :)

Thank you for your understanding.

Patch-set: 3