Update patch set 9
Patch Set 9:

I'm aware the new favicon differs from the current one, but I think it's even better; certainly the 16px version is much more visible than the current one.

That said, I think there's still a small difference between the 32px and 48px layers; the font in 32px is bolder than the one in 48px -- it's best visible if you compare the letter 'e' in the two layers, and also the letter 'n' is bolder in the 32px layer than in the 48px one.

Patch-set: 9
Label: Verified=0