Update patch set 1
Patch Set 1:

As odder says, this happened before 015f5b7, i.e. in 2010 when all the projects were pointed to the new logo on Commons. https://commons.wikimedia.org/wiki/Wikipedia/2.0
FWIW, the current logo on the wiki overlaps with the sidebar and the change to the v2 logo on Commons *was* discussed with the community, see https://fa.wikipedia.org/?diff=prev&oldid=3275529 .
It would be nice to know what's the problem with the "official" logo on Commons but considering that the community has refused to discuss the matter for over 3.5 years I guess it's a lost battle. :) This change just reflects the status quo more cleanly.

Patch-set: 1
Reviewer: Nemo bis <105@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0