Update patch set 2
Patch Set 2:

Hello there! Please do not include "for GCI 2013" in the commit message. The change itself has nothing to do with GCI, so this should not appear in the commit message.

It would also be nice if you could explain in the commit message what exactly you are fixing. This is especially important when changing a binary file, since a normal "diff" is not available for others to view.

Patch-set: 2
Label: Verified=0