Update patch set 1
Patch Set 1: Code-Review-1

The size of the icon seems to be 48x55, not 48x48.

I also wonder whether we shouldn't have a white background behind the text as we had for the Internal icon. However, since there haven't been specific instructions for this I'm fine with keeping the icon with transparent background.

Patch-set: 1
Label: Code-Review=-1