Update patch set 4
Patch Set 4:

I'm not sure whether it matters the order of the layers. So far all the favicons had the 16x16 icons as first layer, then the 24x24 as second and finally the 48x48 as third. In your case 24x24 is the first followed by 16x16.

Perhaps more importantly, it looks like you could make the icons a bit more sharper working at a pixel level, especially in the 16x16 "w" and the end of the square brackets. However, I will defer to Odder the decision of marking your current version -1 or +1 since he is the expert. :)

In any case this is a good first attempt! Thank you.

Patch-set: 4
Label: Verified=0