Create Chinese Wikivoyage (zhwikivoyage)
This patch adds necessary configuration for the creation
of a Chinese Wikivoyage project.

Following instructions on Wikitech, I'm adding the project
to cirrus.dblist and wikidataclients.dblist; it also gets
MW 1.23wmf8 in wikiversions.dat as this will most probably
be the current version of MW when this patch gets merged
and deployed.

I'm using this opportunity to sort wikidataclients.dblist
alphabetically with the Unix sort utility (looks fine).

For settings inside InitialiseSettings.php see bug request
filled by MF-Warburg and follow-up comments.

Bug: 59077
Change-Id: I24dd4a00e00f1deae16209cecda28a096c0b1ac0