Add templateeditor right, group, and restriction
For test.wikipedia; this change adds the 'templateeditor' restriction
level 'between' the 'autoconfirmed' and 'sysop' levels intended for
appying to high-traffic templates that are currently protected.

This commits creates the group 'templateeditor' that has a right
with the same name, allowing members to edit the protected templates,
and gives that right (as well as the ability to grant and revoke
the group) to the 'sysop' group.

The templateeditor right also includes the tboverride permission.

The changes were made to mirror the configuration of en.wikipedia,
for testing purposes.

Bug: 59084
Change-Id: I52a47be04c6aba89562bec40d2438e9df6b7264e