Update patch set 1
Patch Set 1:

> 1) On this patch: doesn't work due to dependencies.

Does anyone have a specific link to what this is about? The entire point of modularized code is that it should be... modular. If the Translate extension is directly dependent on the ULS extension, that's a bug.

> 2) On the proposal: unacceptable to remove ULS from wikis where it's requested by community, "phase0" is not justified.

You know this isn't true. Performance concerns always trump community consensus. Since forever.

> 3) On the commit message: a cherry-picked example of a single page on a single wiki clearly isn't adequate to the proportion of the change, which affects tens of millions pages on hundreds of projects. You're going to need something way more cogent/explanatory than that.

Disabling this extension affects tens of millions of pages how? Doubling the request size is clearly a major bug. Users are being forced to download nearly half a megabyte to read English-language pages. This issue obviously affects more than a single page.

Patch-set: 1