Move WikimediaIncubator extension call to be after Scribunto
The WikimediaIncubator extension checks to see if a certain page
is prefixed and requires the knowledge of which namespace to enact
such a policy. Because it calls the constant NS_MODULE (which needs
to be first defined by Scribunto), this extension hence needs to be
loaded *after* Scribunto has been loaded.

This change only affects the Wikimedia Incubator.

Change-Id: I00f6cd55c2e8d783cefc2615b5730f0e45fff03a