Resetting legacy channel names on labs and enabling IRC-RC echo again
some wikis have legacy channelnames defined (see gerrit 47307). Using
them on labs resulting in mixing up the RC for prod and beta in one
channel because the lookup based on the wikiname, which is the same for
prod and labs.
Resetting wmgRC2UDPPrefix to let CommonSettings.php determine the
channel name based on the hostname, which does not create any collisions
while using labs hostnames.

As the bug should be fixed, enabling RC2UDP again.

Bug: 60013
Bug: 56758
Change-Id: Idc34dc69e5e91efdf07b7b8312a58069e77cdb06