Update patch set 2
Patch Set 2:

Initial idea adding the item-create permission to the 'createeditmovepage' and everything else for interacting with items to the 'editpage' group and totally ignoring touching properties for now.

The alternative is we split this up further and have wikibase groups as suggested in the bug.

Comments?

Patch-set: 2
Reviewer: Addshore <106@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0