Update patch set 2
Patch Set 2:

I think we should ditch these fine grained permissions. we are not using them. the item-* stuff kind of makes sense (but we'd need the same for property-* too), but i don't see why we need multiple permissions for labels, aliases, etc. maybe "edit-terms", but that's it. Maybe something for labels of properties?

We should really discuss and decide this.

Patch-set: 2