Update patch set 4
Patch Set 4:

Can you please give precise instructions to Yatin to abandon that patch? Is it enough to abandon it from the Gerrit interface or should this patch here be rebased as well?

As far as I remember, Yatin worked really hard to get through all the Git & Gerrit hops in his Windows system. If the current dependency can be solved via web interface alone that would be great.

Patch-set: 4