Split some wikis into more shards
The goal is to keep the shards under 2gb where possible but to have less
than 20 each.  20 is pretty arbitrary at this point but 2gb helps with
maintenance.  If the shards are too big they are a pain to move around.

DEPLOYMENT: This won't take effect until the modified wikis are inplace
reindexed but this isn't important enough for us to do one unless we have
something else coming.

Change-Id: I15d6a8da6d2a879c624f5ce01caa906c704a74b0