Revert "Replace easter egg by a more explaining message"
As well as follow up "Fix test broken by Idda8cff80".

For years and years we have build trust among our community by reaching
consensus on change.  Handling review and merging change on a Sunday
despite minus votes is not friendly.

Additionally that patch touch a very sensible part of code which need to
be carefully reviewed for security issues.  Even if it look harmless, it
can lead to leak of private data.

This reverts commits:
    f4f834ba0ffdc653e9219035be46a3fffa6a5367
    a37d2b42299a7fab75e6dd786374c2ae65d6ceeb

Change-Id: I016718c8cb55b250d4cb535141069a6e553528e6