Update patch set 1
Patch Set 1:

As someone who has reported it as a security bug in the past, I object to the term "easter egg".  An easter egg is something that makes the finder happy of his achievement and not the object of ridicule.

If someone likes to laugh, there's plenty of that on YouTube, Dailymotion and whatnot.

Patch-set: 1
Reviewer: Tim Landscheidt <216@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0