Update patch set 1
Patch Set 1:

Can't tell if you did, but did you run createTxtFileSymlinks.sh? If that makes the same change then that's good. If it makes other changes, perhaps this commit should be a general noc sync fixup? Unless it makes bad changes, in which case this change is fine and we can fix createTxtFileSymlinks.sh later.

Patch-set: 1
Label: Code-Review=0