Update patch set 1
Patch Set 1:

This was split off from VectorBeta for code separation concerns, and is going to be deployed in a couple of weeks (once the ongoing design review is completed). prtksxna who is the primary author of that extension was recently hired by the design team just for this. And in the last few hours a lot of the code has been merged, and having this on betalabs would be rather useful, I think. Besides, if this were to target deployment in about 2 weeks, this needs to go to betalabs now.

Patch-set: 1
Reviewer: Yuvipanda <168@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0