Remove ArticleFeedbackv5 from Wikimedia wikis
TL;DR: it's not functional and there are no resources to fix it, see bug 54197.

Per Fabrice, two weeks ago it was time for a "resolution next week".
<http://lists.wikimedia.org/pipermail/ee/2014-January/000882.html>
The plan was to discontinue AFT in case it gained no support from one or
two big wikis: it was semi-disabled on en.wiki, disabled on de.wiki,
rejected on fr.wiki.
<https://www.mediawiki.org/wiki/Article_feedback/Version_5/Release_Plan_2013>

In particular, fr.wiki has been waiting for disabling for several weeks,
after the pilot ended in December 2013.
<https://fr.wikipedia.org/wiki/Wikip%C3%A9dia:Sondage/Outils_d%27%C3%A9valuation>
<https://fr.wikipedia.org/wiki/Wikip%C3%A9dia:Le_Bistro/22_janvier_2014#Outil_de_retour_des_lecteurs>
<https://fr.wikipedia.org/wiki/Wikip%C3%A9dia:Le_Bistro/25_janvier_2014#Retrait_complet_.28.2B_date.29>

A more detailed report/recap by Fabrice will follow, for thorough
information of stakeholders not as directly involved in decision, like
wikis where AFTv5 was not enabled yet but which were curious.

Bug: 61163
Change-Id: Ief22aa5fa6ec19dbe368f79392c74a507945e02a