Remove unused third/fourth parameters for wikiversions
This patch makes multiversion scripts only parse the first two space-separated
fields in wikiversions.dat. The third field (external version) and fourth
"field" (an optional comment, preceded by '#') are discarded. To ease
migration, the structure of CDB files is preserved, with the external version
field hard-coded to the empty string.

Change-Id: Id87a90474ffc1fce6eed58c228fc91fcd36e5dbb