Update patch set 3
Patch Set 3: Code-Review-1

Hi guys,

I would appreciate if we held back for a few weeks on merging this patch.

As discussed with Nemo on the editor engagement list, we would like to give editors a couple weeks to move the most useful feedback to their article talk pages.

I have posted a report on AFT5 here on MediaWiki.org and am about to announce the proposed removal of AFT5 in a couple weeks:
https://www.mediawiki.org/wiki/Article_feedback/Version_5/Report

Please honor this reasonable request, so that editors who benefit from the tool can take the appropriate action with sufficient notice.

Thanks for your understanding!


Fabrice

Patch-set: 3
Label: Code-Review=-1