Update patch set 5
Patch Set 5: Code-Review-1

As far as I know nothing like that has been told to those communities who expressed interest (it certainly hasn't been told to me when I did), but if you know otherwise, feel free to point out where it happened.

As for being unmaintained, the status of the extension is  stable according to mediawiki.org. What sort of maintenance is expected? There are lots of extensions enabled on some WMF wiki or other which have not had a significant code change for years and the original maintainer is probably long inactive. Certainly you want there to be an active maintainer when something is added to the WMF cluster, in case anything goes wrong, but AFTv5 has been added a long ago, has been running without big changes and without causing any technical problems for a good while, and it does directly alter any major feature like article editing or storage, so it can be very easily disabled in case it becomes incompatible at some point, or if any major bug is discovered. So as far as I see the extension easily meets the conditions for leaving an already installed extension enabled; and other extensions in similar situations are not disabled, certainly not with the urgency you are displaying here.

Anyway, I don't think a gerrit commit is the right place to have this discussion. Unless there is some sort of emergency, we should take this to wikitech-l.

Patch-set: 5