Add templateeditor user group, protection to rowiki
This patch adds the 'templateeditor' restriction
level between the 'autoconfirmed' and 'sysop' levels
intended for high-traffic templates.

It also creates the 'templateeditor' user group that
has a right of the same name, allowing members to edit
protected templates, and gives that right (as well as
the ability to grant and revoke the group) to the
'sysop' user group.

See bug for URL with on-wiki consensus.

Bug: 61172
Change-Id: Ib6c61cb7cd672f24d115b2cdcd356a607bda1f2e