Add qa_automation group and grant it Flow rights
Browser tests for Flow test the deletion and suppression of posts and
topics.  These operations require permissions normally granted to admins
and oversighters.  We don't want to give Selenium_user those rights, so
on test2wiki add a qa_automation group and give it flow-delete and
flow-suppress permissions.

An admin would add Selenium_user to this qa_automation group on
test2wiki and tests should pass.

Bug: 58375
Change-Id: I7523c867b1237c5b046eaa5b8954fc6d34188bdc