Update patch set 1
Patch Set 1:

(1 comment)

-2 is inappropriate in this case.

Even if the patch is totally wrong, the intended result of the patch is correct and so should be -1 not -2. (-2 would be e.g. if someone submitted a patch to give every user a CAPTCHA on every edit. even if the code changes were correct, it's a change we don't want)

from me:
-1 for wrong IP and wrong array key. ("IP" vs. "range") (see inline comments)

Patch-set: 1
Reviewer: Jeremyb <48@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0