Enable lower-res .ogv transcodes for fallback player on slow machines
Currently for high-res source files we only produce a 480p version
in Ogg Theora; on slower machines with non-native software decoding
fallbacks (currently a Java applet, in future may be JavaScript or
even Flash for some machines) this is just too many pixels to push.

There are already definitions ready to go for 360p and 160p .ogv
output; this config patch enables them so they'll get built in future.

Bug: 61690
Change-Id: Ib2dd92626826ebeef71e0b5f77d5c5f2f67f80a1