Update patch set 1
Patch Set 1: Code-Review-1

"Meta clearly doesn't want Flow just yet" - well, no, Odder and MZMcBride do not want Flow to be enabled on Meta. That is not a community discussion or anything approaching consensus.

To be 100% crystal clear, Flow is not being forced on anyone who doesn't want to use it - we put it on a test page so we could test it to make sure it works for the Meta users who *do* want to use it on Meta: the Program Evaluations team, who just want to use it for feedback on their learning modules pages. 

I'm having a really hard time seeing why Odder and MZ feel so strongly that those users *can't* have Flow on their talk page. Why do 2 people who won't even be looking at or using the software get to dictate whether people who want it do or don't get it?

Patch-set: 1
Reviewer: Maryana <455@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1