Update patch set 1
Patch Set 1:

But I can turn this argument on its head and say: why do two people (you and Erik) get to decide what extensions does a whole community get to use or not?

You enabled Flow on Meta without asking anyone about their opinion on the subject; it's a fact. There was no community consensus to enable it in the first place, and I kindly ask you not to distort the situation.

As for your other argument, /you/ do not get to decide if I can speak my mind on a configuration change or not. 

I have always thought that the opinion of volunteer users was of the same value to the movement as those of paid staff, but it appears I was wrong.

If you want to test Flow, you not only have a whole wiki creates specifically for this purpose, but also a whole Beta cluster. No one will complain if you test your software there.

Patch-set: 1