Update patch set 1
Patch Set 1:

"As for your other argument, /you/ do not get to decide if I can speak my mind on a configuration change or not."

Has someone censored/edited/redacted your comments here, bugzilla, wiki, or the mailing list?

I'm not party to all the discussions going on surrounded this, but if this is happening and someone at the WMF isn't allowing you to speak or somehow prevented you from submitting this change for review, I'd like to know about that because that's a serious problem if so and I'll be your staunchest advocate in that regard.

Patch-set: 1