Update patch set 1
Patch Set 1:

"EventLogging is hardly a good comparison because it is utterly un-user facing (There's also a difference between enabling things by default on all wikis, and enabling it on couple wikis)."

Brian, please take the time to review the original patch enabling Flow on Meta.

1) No Flow code is run/executed on pages Flow is enabled on
2) Flow is only enabled on a single test page which is not visible
3) Flow was requested by a community within Meta

EventLogging had more of a user-facing impact (performance) than Flow does in its current state. I personally felt that this user-facing impact was positive because it meant the eventual removal of the highly-unstable click-tracking, but I don't recall asking anyone's permission from the community in this regard.

…

Even though I'm the Director of Features Engineering and Flow reports directly to me, I was not consulted, nor was I asked permission on this deploy. That's how it should be. I assume good faith in our product managers on product decisions, and my engineers on engineering ones, and don't second guess unless I'm privy to some knowledge they don't have. I trust you on multimedia engineering decisions in the same manner, and I don't +1/-1 your stuff unless I have something to add. I assume the same good faith with Odder here, even though he and I are in disagreement. That's a healthy disagreement.

When you +1'd this change, what's your reasoning? Is there something my engineering team doesn't know and should about Flow that's unique to Meta? Have you been following and participating in the discussion of Flow on Meta closer than Maryana?

…
"Usually the burden of consensus falls with the people wanting the extension enabled. Most times such consensus requirements have been bypassed in the past from what I remember, especially for wikis that are not named enwiki, it resulted in a lot of angry people."

I agree that pre-deploy consensus would be idea ignoring any costs or the logistics of such.

However, there is no "burden of consensus" nor has there ever been. Besides, such a burden flies in the face of being bold. The actual decision-making seems to me much more nuanced on behalf of product and community (again, I might emphasize I restrict my decisions to engineering and engineering-related staffing). Though if I can be so bold as to venture an opinion, the decision-making seems to be highly centered around impact and deploying scarce resources when potential negative impact is greatest.

Here are some examples from my perspective (so there's a lot wrong because I'm not in product or community):

Echo could not go out under limited testing in a wiki. The product manager and community liaison tried to get consensus (or actively informed the community) before the WMF deployed it in most cases. The WMF didn't get consensus in a few cases for early deploys before deploying it anyway (mediawiki and enwiki being the two that I remember, possibly also Meta). Bugs were prioritized when the impact was negative and rollback was considered (c.f. Orange Bar of Death), but rollback didn't occur, and eventually compromise was reached. Eventually Echo was stable and popular enough after widespread testing on many of the most high-traffic wikis, that most community complaints centered on the slow pace of rollouts to further wikis. The logistics of reaching out to every wiki became untenable. At that point, the metric was reversed from reaching-out pre-launch to announcing projected launch dates and allowing communities to gather consensus to opt out. At that point only two wikis took the opted-out, but eventually changed their feeling of the matter (cost-benefit weighed heavily and/or Echo improvements were made based on their concerns).

ArticleFeedbackToolv5, unlike Echo, could be deployed on a limited subset of articles. However, monitoring AFTv5 free form comment field for abuse could have site-wide negative impact. Because of this, the WMF reached out to some enwiki admins before deploying (and similarly in every wiki the WMF deployed a test to). Eventually, I believe, all the wikis that tried it (I think there were 3) determined the spam-helpful ratio just wasn't worth it and it was rolled back and the test cancelled. Having said this, the WMF never asked permission from the larger community of those wikis before deploying, but the WMF did ask the people most affected by it over the course of years to participate, even if in the end it didn't work out. There are some of the sister-projects that wish to trial/deploy AFTv5, but WMF resource constraints prevent that as there is no support infrastructure. :-( I will guess that product will consider revisiting AFT iterations only if better infrastructure to absorb/manage abuse is in place. This is a case where product would have to follow engineering and core work for anything to succeed.

NewPages can work in parallel to the existing infrastructure. After a lot of objection from some in the community, and support of others, it was deployed. Eventually it did well, and a number of other language wikis requested the feature. However the original scope in response to a rejected-RfC unique to enwiki, so it was built/hardcoded with a lot of processes unique to that wiki. As such, scarcity prevents it from being re-engineered for any other wiki. WMF-resource constraints mean that it is barely functional on enwiki, though my hope is that Workflows may allow the WMF to revisit it and rebuild the extension so that wikis can self-serve and maintain the extension.

I don't believe the WMF ever asked permission of the wider wiki community before deploying post-edit feedback, guided tours, or the new sign-in or on boarding experiments, of which there were many iterations and all of which were user facing. However, the WMF did actively try to inform the community members and when community members (and developers) reached out all these products had some changes. I can think of a number of issues brought up from the font choice, the code and style used to show dialog boxes, to wording and formatting that were resolved based on community input/suggestions.

I'm sure the CL's (of which there was only one at the time) and the PM's could have done a better job of reaching out in all these cases (nobody is perfect), but the logistics of such is constrained by resources on the WMF side and an informal, decentralized, consensus-building framework on the wiki communities. :-( Ideally consensus is best, but when people reach the point of a poor approximation of such (as is often the case), unless I have some special knowledge (it's rare, but it happens :-D), I assume good faith in my fellow peers and community members, as I do in you, and the other participants here.

This is one person's perspective and I hope it added something to the discussion. Even if I had +2/-2, I'd have only +1/-1'd this change since there is no engineering reason to add or remove this change.

Patch-set: 1