Update patch set 1
Patch Set 1: Code-Review-1

Looking at it; I actually agree with the -1's so far. A community consensus in my opinion is necessary for a forced change affecting the whole community. This change form what I see is 1) optional and 2) not forced. This has happened before when the WMF have strictly gone 'We are doing this, simple' and we accept it simply because if the WMF are doing a Foundation-initiative which in my opinion is similar to a Government-initiative. We can override it (and I am sure the WMF mainly Maryana since she is the product's lead) if there is strong consensus against this and not a 'You did this without asking us - so we'll do this without asking you' case.

TL;DR - This is a Foundation-initiative and the community can override with WMF agreeing if we get consensus not a futile commit war over who is right. A small small percent of the community or the WMF for wanting to test and provide an optional feature. If of course this turns into a 'We will force this on you now' I'd agree for it's removal but it does not seem like this one bit.

Patch-set: 1
Reviewer: John F. Lewis <589@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1