Update patch set 1
Patch Set 1:

To clarify, my +1 is meant as symbolic support for social reasons. I have no technical objections to deploying flow to wikis for testing purposes.

I personally think its disrespectful to the communities in question to deploy new extensions without asking first (mediawikiwiki being an exception as it is "dev" owned). Particularly when its things that affect workflow or are easily user noticable. Even more so when the software is experimental like Flow is. I believe wikis should be "owned" by their respective communities and special treatment of a specific wiki should require that wikis approval. I recognize that logistically it can be hard to get consensus, but in this case it doesnt seem like it would have been very difficult to ask. At the very least, if not consensus, a notice on meta saying we are doing X unless someone objects could have been given. And now that there are real people at meta (mz and Odder) who clearly dont want it without discussion, it seems like it would be trivial to just revert pending the establishment of consensus.

Patch-set: 1