Update patch set 1
Patch Set 1:

We do not operate on the basis of obtaining per-wiki consensus prior to any software change to Wikimedia Foundation wikis. We never have, and we never will. It'd be an absurd way to operate, it would be impossible to get anything done, ever. We do strive to communicate, solve problems real users have, and resolve conflicts, especially if there are major points of contention. What are the material issues with the deploy at this time?

Patch-set: 1