Update patch set 1
Patch Set 1:

To be clear, if we failed to communicate effectively and well about what was about to happen (which I understand to be a very limited test deployment), that's our bad and we should do better - sorry. But if we're going to first principles, I want to be very clear about what we will do and what we won't do.

Patch-set: 1
Reviewer: Eloquence <20@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0