Update patch set 1
Patch Set 1:

@PiRSquared17	 It was a little bit tongue-in-cheek. I was just implying the logistics of "gathering community consensus to enable Flow" on Meta, given its unique position and mission, multitude of focuses and interests, both inclusive and decentralized organization, is… impractical. Theoretically, the WMF would have to get consensus on EVERY wiki before there'd said to have been consensus on meta. :-)

I apologize if it the comment could be construed to denigrate the Meta community, because I actually find it a very interesting one, and… how do the kids say? … "so… meta"? :-D

To the extent that the members of the Meta community actually impacted by the test page on Flow, it was my understand they were consulted as I believe this was a request that actually came from the people affected.

Patch-set: 1