Update patch set 1
Patch Set 1:

I thought we decided to use Bugzilla for discussion, not Gerrit. I've replied on bug 61729.

While it would be nice to cleanly disable this extension until there's local community consensus to deploy alpha discussion software to Meta-Wiki, it's easy enough, if not a bit dirty, to disable the use of Flow on Meta-Wiki.

I continue to recommend discussion and collaboration with the Meta-Wiki community rather than simply trying to steamroll it. I think the course of action being taken by another of Terry's teams will only engender more ill will toward this project and those to come.

Patch-set: 1
Label: Code-Review=0