Update patch set 2
Patch Set 2: Code-Review+1

> Will the network latency coming from the writes to the other DC slow down beta (significantly)?

Valid concern, but the best way to find out is to give it a shot and see.

Hashar: voting +1 but consider it a +2; leaving it for you to merge at a time when you can monitor the outcome.

Patch-set: 2
Reviewer: Ori.livneh <255@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1