beta: disable memcached accross datacenters
Chris McMahon reported a bunch of slowness in browser tests. That is
most probably due to instances writing to memcached in another
datacenter in addition to the local DC.

That more or less revert 4876db8 which introduced multiwrite on beta.
We do not need to populate eqiad cache, whenever we migrate the cache
will be cold and session dropped, not a big deal on beta.

Change-Id: I06d07cc3e6ce2ebbf474799f1ed24623d15e3cb9