Update patch set 1
Patch Set 1: Code-Review-1

I know WHOIS says 186.193.0/20 but is that valid syntax for whatever will be parsing the range? I'd use 186.193.0.0/20 to be safe.

Also, this is a bigger than normal range. AIUI, it's not a limit of 100 for the range but rather 100 per IP in the range. range size is ~4k IPs.

Patch-set: 1
Reviewer: Jeremyb <48@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1