Update patch set 1
Patch Set 1:

To clarify my earlier reservation, I'll explain that enabling webfonts means any page on the wiki can in potential load any number of fonts. I would prefer to be consistent and predictable and load one font always; this would also help ensure that it is cached for the maximum number of users.

I am wonder if this problem could be solved by adding the following line to the top of <he.wikisource.org/wiki/MediaWiki:Common.css>:

    @import url(http://fonts.googleapis.com/earlyaccess/alefhebrew.css);

(As I explained on IRC, I have no special love for Google; we can have this point to our own copy of the fonts instead.)

Is there a reason why this approach would be inferior? If so, I am entirely amenable to supporting the redeployment of web fonts of hewikisource. But please help me understand the full picture.

Patch-set: 1