Increase the network performance sampling rate
Since Media Viewer is a beta feature at the moment, we're
collecting too little data at the current sampling rate to make
truly informed decisions about performance. This patch ups the
sampling rate to 1:10 rather than 1:1000, and sets it to go up by
a power of ten for every week today, to ensure it is not left on
by default after being enabled for all users.

Change-Id: Ib539f96eb79707549c31cef6898920e0340e41f1