Beta cluster MemcachedPeclBagOStuff: use PHP serialization
The PECL memcache client stores the serialization format of each cached value
in the flags bitfield that accompanies each value. When it gets an item, the
client will use the value of that flag to determine which decoder to use. In
other words, the serializer option only determines how newly-set values are
encoded.

HHVM does not support igbinary serialization, and Aaron's casual testing
indicated that the savings provided by igbinary were minimal. So it makes sense
to switch back to PHP serialization, which is supported by HHVM.

There should be no substantial impact on the beta cluster: all the keys that
are currently in the cache will continue to be readable.

Change-Id: I46a9d180bfa5f954a7c793e6fdb8b1c7b19cca3a